requirements.txt
Flask==3.0.0
opencv-python-headless==4.8.1.78
numpy<2.0.0
pymongo==4.6.1
python-dotenv==1.0.0